Can now ask a server for its player count
... using the `get_player_count_of_server` function. Also moved the `get_servers` and `create_server` functions to a new file called `functions.py` which also stores the new `get_player_count_of_server` function.
